Refactor upload in separate feature
